mcp/pyproject.toml
[project]
name = "irpapers-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.5.0",
    "weaviate-client[agents]>=4.19.2",
]

[tool.setuptools]
packages = ["src"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
